% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_Checker.R
\name{bib_Checker}
\alias{bib_Checker}
\title{bib_Checker function}
\usage{
bib_Checker(Rmd_file = "RmdFileName", Bib_file = "BibFileName")
}
\arguments{
\item{Rmd_file}{file name of R Markdown file}

\item{Bib_file}{file name of Bib file}
}
\description{
bib_Checker function
}
\examples{
# bib_Checker(Rmd_file = "RmdFileName",Bib_file = "BibFileName")
}
