% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib_to_df.R
\name{bib_to_DF}
\alias{bib_to_DF}
\title{bib_to_DF}
\usage{
bib_to_DF(Rmd_file, Bib_file, list_ampersand = F, cite_ampersand = F)
}
\arguments{
\item{Rmd_file}{file name of R Markdown file}

\item{Bib_file}{file name of Bib file}

\item{list_ampersand}{combinle the last author with & or not in Bibliography}

\item{cite_ampersand}{combinle the last author with & or not in in-line citation}
}
\value{
Prepare for citation database
}
\description{
bib_to_DF
}
\examples{
# bib_to_DF(Rmd_file = "RmdFileName",Bib_file = "BibFileName")
}
