% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_citation.R
\name{inLineCitation}
\alias{inLineCitation}
\title{inLine Citation}
\usage{
inLineCitation(st, bib.df)
}
\arguments{
\item{st}{A line of text}

\item{bib.df}{Bib.df object created by bib_to_DF function}
}
\value{
replace statement
}
\description{
inLine Citation
}
\examples{
# bib_to_DF(Rmd_file = "RmdFileName",Bib_file = "BibFileName")
}
