% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{inLineCite_ENG}
\alias{inLineCite_ENG}
\title{in-Line Cittion(in English)}
\usage{
inLineCite_ENG(df)
}
\arguments{
\item{df}{Bib.df File from jpa_cite}
}
\description{
in-Line Cittion(in English)
}
