% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{inLineCite_JPN}
\alias{inLineCite_JPN}
\title{in-Line Cittion(in Japanese)}
\usage{
inLineCite_JPN(df)
}
\arguments{
\item{df}{Bib.df File from jpa_cite}
}
\description{
in-Line Cittion(in Japanese)
}
