% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{inLineCite_TR}
\alias{inLineCite_TR}
\title{in-Line Cittion(Japanese Transration book)}
\usage{
inLineCite_TR(df, ampersand)
}
\arguments{
\item{df}{Bib.df File from jpa_cite}
}
\description{
in-Line Cittion(Japanese Transration book)
}
