% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_rmd.R
\name{jjbct}
\alias{jjbct}
\title{R Markdown template for paper of Japanese Journal of Behavioral and Cognitive Therapies}
\usage{
jjbct()
}
\description{
R Markdown template for paper of Japanese Journal of Behavioral and Cognitive Therapies
}
