% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_cite.R
\name{jpa_cite}
\alias{jpa_cite}
\title{Add citation with JPA format}
\usage{
jpa_cite(Rmd_file, Bib_file, use.et.al.first = TRUE)
}
\arguments{
\item{Rmd_file}{file name of R Markdown file}

\item{Bib_file}{file name of Bib file}
}
\value{
Make reference list and add it to R Markdown file
}
\description{
Add citation with JPA format
}
\examples{
# jpa_cite(Rmd_file = "RmdFileName",Bib_file = "BibFileName")
}
