% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_rmd.R
\name{jpa_jjp2}
\alias{jpa_jjp2}
\title{R Markdown template for paper of the Japanese Journal of Psychology ver2}
\usage{
jpa_jjp2()
}
\description{
R Markdown template for paper of the Japanese Journal of Psychology ver2
}
