% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_cite.R
\name{jpr_cite}
\alias{jpr_cite}
\title{Add citation with JPR format}
\usage{
jpr_cite(Rmd_file, Bib_file)
}
\arguments{
\item{Rmd_file}{file name of R Markdown file}

\item{Bib_file}{file name of Bib file}
}
\value{
Make reference list and add it to R Markdown file
}
\description{
Add citation with JPR format
}
\examples{
# jpr_cite(Rmd_file = "RmdFileName",Bib_file = "BibFileName")
}
