% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_reply.R
\name{make_correction_table}
\alias{make_correction_table}
\title{Make crrection table}
\usage{
make_correction_table(old_manuscript, new_manuscript)
}
\arguments{
\item{old_manuscript}{vector of old content of manuscript}

\item{new_manuscript}{vector of new content of manuscript}
}
\description{
Make crrection table
}
