% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_reply.R
\name{new_m}
\alias{new_m}
\title{Add new content of manuscript to crrection table}
\usage{
new_m(new_text)
}
\arguments{
\item{new_text}{new content of manuscript}
}
\description{
Add new content of manuscript to crrection table
}
