% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_reply.R
\name{old_m}
\alias{old_m}
\title{Add old content of manuscript to crrection table}
\usage{
old_m(old_text)
}
\arguments{
\item{old_text}{old content of manuscript}
}
\description{
Add old content of manuscript to crrection table
}
