% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{print_EName}
\alias{print_EName}
\title{Print name function(English)}
\usage{
print_EName(st, ampersand = T, switchFLG = FALSE)
}
\arguments{
\item{st}{Strings of name}

\item{ampersand}{it TRUE, combine last author with ampersand else "and"}

\item{switchFLG}{switch the order of first name and last name}
}
\description{
Print name function(English)
}
