% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{print_English_article}
\alias{print_English_article}
\title{Print bib info function(English article)}
\usage{
print_English_article(df, underline = F)
}
\arguments{
\item{df}{Strings of Bib info}

\item{underline}{check flag for underline}
}
\description{
Print bib info function(English article)
}
