% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{print_English_book}
\alias{print_English_book}
\title{Print bib info function(English book)}
\usage{
print_English_book(df, underline = F)
}
\arguments{
\item{df}{Strings of Bib info}
}
\description{
Print bib info function(English book)
}
