% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{print_English_inbook}
\alias{print_English_inbook}
\title{Print bib info function(in English inbook)}
\usage{
print_English_inbook(df)
}
\arguments{
\item{df}{Strings of Bib info}
}
\description{
Print bib info function(in English inbook)
}
