% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{print_English_inproceedings}
\alias{print_English_inproceedings}
\title{Print bib info function(in English Proceedings)}
\usage{
print_English_inproceedings(df)
}
\arguments{
\item{df}{Strings of Bib info}
}
\description{
Print bib info function(in English Proceedings)
}
