% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{print_exceptional_bib}
\alias{print_exceptional_bib}
\title{Print bib info function(Original Style references)}
\usage{
print_exceptional_bib(df)
}
\arguments{
\item{df}{Strings of Bib info}
}
\description{
Print bib info function(Original Style references)
}
