% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{print_incollection}
\alias{print_incollection}
\title{Print bib info function(in collection)}
\usage{
print_incollection(df)
}
\arguments{
\item{df}{Strings of Bib info}
}
\description{
Print bib info function(in collection)
}
