% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{print_Japanese_article}
\alias{print_Japanese_article}
\title{Print bib info function(Jaopanese article)}
\usage{
print_Japanese_article(df)
}
\arguments{
\item{df}{Strings of Bib info}
}
\description{
Print bib info function(Jaopanese article)
}
