% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{print_Japanese_inbook}
\alias{print_Japanese_inbook}
\title{Print bib info function(in Japanese inbook)}
\usage{
print_Japanese_inbook(df)
}
\arguments{
\item{df}{Strings of Bib info}
}
\description{
Print bib info function(in Japanese inbook)
}
