% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_bib_functions.R
\name{print_Japanese_inproceedings}
\alias{print_Japanese_inproceedings}
\title{Print bib info function(in Japanese Proceedings)}
\usage{
print_Japanese_inproceedings(df)
}
\arguments{
\item{df}{Strings of Bib info}
}
\description{
Print bib info function(in Japanese Proceedings)
}
