% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_rmd.R
\name{render_jjbct}
\alias{render_jjbct}
\title{render function for paper of Japanese Journal of Behavioral and Cognitive Therapies}
\usage{
render_jjbct(Rmd_file = "Untitled.Rmd", Bib_file = "reference.bib")
}
\arguments{
\item{Rmd_file}{file name of R Markdown file}

\item{Bib_file}{file name of Bib file}
}
\description{
render function for paper of Japanese Journal of Behavioral and Cognitive Therapies
}
