% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_rmd.R
\name{render_jpr}
\alias{render_jpr}
\title{render function for paper of Japanese Psychological Review}
\usage{
render_jpr(Rmd_file, Bib_file)
}
\arguments{
\item{Rmd_file}{file name of R Markdown file}

\item{Bib_file}{file name of Bib file}
}
\description{
render function for paper of Japanese Psychological Review
}
