% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_rc.R
\name{set_rc_jabct}
\alias{set_rc_jabct}
\title{Set Research Compendium for JABCT}
\usage{
set_rc_jabct(file_name = "paper")
}
\arguments{
\item{file_name}{file name and directory name of RMarkdown}
}
\value{
Make directories for Research Compendium of JABCT and R Markdown file
}
\description{
Set Research Compendium for JABCT
}
\examples{
# set_rc_jabct("rmarkdown_for_reproducibility")
}
