% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa.R
\name{set_rc_jpa}
\alias{set_rc_jpa}
\title{Set Research Compendium for JPA}
\usage{
set_rc_jpa(file_name = "paper")
}
\arguments{
\item{file_name}{file name and directory name of RMarkdown}
}
\value{
Make directories for Research Compendium of JPA and R Markdown file
}
\description{
Set Research Compendium for JPA
}
\examples{
# set_rc_jpa("rmarkdown_for_reproducibility")
}
