% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpa_cite.R
\name{jpa_cite}
\alias{jpa_cite}
\title{Add citation function}
\usage{
jpa_cite(file_name)
}
\arguments{
\item{file_name}{file name of R Markdown file}
}
\value{
Make reference list and add it to R Markdown file
}
\description{
Add citation function
}
\examples{
# jpa_cite("template.Rmd")
}
